+++ bin/src/connection/Connection.d
@@ -1,6 +1,0 @@
-src/connection/Connection.o: ../src/connection/Connection.cpp \
- ../src/connection/Connection.h ../src/connection/ConnectionException.h
-
-../src/connection/Connection.h:
-
-../src/connection/ConnectionException.h: