+++ bin/src/connection/ConnectionException.d
@@ -1,5 +1,0 @@
-src/connection/ConnectionException.o: \
- ../src/connection/ConnectionException.cpp \
- ../src/connection/ConnectionException.h
-
-../src/connection/ConnectionException.h: