+++ bin/src/connection/InputConnection.d
@@ -1,9 +1,0 @@
-src/connection/InputConnection.o: ../src/connection/InputConnection.cpp \
- ../src/connection/InputConnection.h ../src/connection/Connection.h \
- ../src/connection/ConnectionException.h
-
-../src/connection/InputConnection.h:
-
-../src/connection/Connection.h:
-
-../src/connection/ConnectionException.h: