+++ bin/src/connection/OutputConnection.d
@@ -1,9 +1,0 @@
-src/connection/OutputConnection.o: ../src/connection/OutputConnection.cpp \
- ../src/connection/OutputConnection.h ../src/connection/Connection.h \
- ../src/connection/ConnectionException.h
-
-../src/connection/OutputConnection.h:
-
-../src/connection/Connection.h:
-
-../src/connection/ConnectionException.h: